requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.20
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
python-multipart>=0.0.6